requirements.txt
yfinance>=1.1.0
curl_cffi>=0.7.0
pandas==2.2.0
numpy==1.26.4
python-dotenv==1.0.1
python-dateutil==2.9.0
apache-airflow==2.8.0
apache-airflow-providers-postgres==5.10.1
apache-airflow-providers-http==4.10.1
pyspark==4.1.1
pyarrow==15.0.0
pytest==8.0.0
